requirements.txt
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cryptography==41.0.3
defusedxml==0.7.1
deprecation==2.1.0
dj-database-url==2.1.0
Django==4.2.4
django-allauth==0.56.1
django-ckeditor==6.7.0
django-js-asset==2.1.0
django-meta==2.4.2
django-robots==6.1
djangorestframework==3.15.2
gotrue==2.11.0
gunicorn==21.2.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.27.2
hyperframe==6.0.1
idna==3.4
oauthlib==3.2.2
packaging==23.1
Pillow==10.0.0
postgrest==0.16.11
psycopg2-binary==2.9.10
pycparser==2.21
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
realtime==1.0.3
requests==2.31.0
requests-oauthlib==1.3.1
sentry-sdk==2.19.2
six==1.17.0
sniffio==1.3.1
sqlparse==0.4.4
storage3==0.7.7
StrEnum==0.4.15
supabase==2.6.0
supafunc==0.5.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
uvicorn==0.32.1
websockets==12.0
whitenoise==6.5.0
